pyproject.toml
[project]
name = "prat-o-mat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "pyannote-audio>=4.0.2",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
]
